+++ src/dns_exporter/collector.py
@@ -63,24 +63,15 @@ class DNSCollector(Collector):
                 addr=self.config.proxy.hostname,
                 port=self.config.proxy.port,
             )
-            # use proxy socket for protocols udp, tcp, doh
             dns.query.socket_factory = socks.socksocket
-            # for protocol doq - method depends on the dnspython version
             # https://github.com/rthalley/dnspython/issues/1059
             if hasattr(dns.quic._sync, "socket_factory"):
                 dns.quic._sync.socket_factory = socks.socksocket
-            else:
-                socket.socket = socks.socksocket  # type: ignore
             logger.debug(f"Using proxy {self.config.proxy.geturl()}")
         else:
-            # restore regular socket.socket for protocols udp, tcp, doh
             dns.query.socket_factory = socket.socket
-            # for protocol doq - method depends on the dnspython version
-            # https://github.com/rthalley/dnspython/issues/1059
             if hasattr(dns.quic._sync, "socket_factory"):
                 dns.quic._sync.socket_factory = socket.socket
-            else:
-                socket.socket = _socket  # type: ignore
             logger.debug("Not using a proxy for this request")
 
     def describe(self) -> Iterator[Union[CounterMetricFamily, GaugeMetricFamily]]: